split out tests on the offer cache
